requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.2.0
cryptography
pymysql
PyQt5 requests
